+++ aiohomematic/const.py
@@ -930,6 +930,40 @@ _IGNORE_ON_INITIAL_LOAD_PARAMETERS: Final[frozenset[Parameter]] = frozenset(
 )
 
 
+_CLIMATE_TRANSMITTER_RE: Final = re.compile(r"CLIMATE.*(?:TRANSMITTER|TRANSCEIVER)")
+
+
+def channel_is_transmitter(channel_type_name: str) -> tuple[DataPointCategory, ...]:
+    """Reck if a channel type name matches common wildcard patterns."""
+    result: list[DataPointCategory] = []
+    if bool(_CLIMATE_TRANSMITTER_RE.search(channel_type_name)) is True:
+        result.append(DataPointCategory.CLIMATE)
+    return tuple(result)
+
+
+_CLIMATE_RECEIVER_RE: Final = re.compile(r"CLIMATE.*(?:TRANSCEIVER|RECEIVER)")
+_CLIMATE_SWITCH_RECEIVER_CHANNELS: Final = ("SWITCH_VIRTUAL_RECEIVER",)
+
+
+def channel_is_receiver(channel_type_name: str) -> tuple[DataPointCategory, ...]:
+    """Return if a channel type name matches common wildcard patterns."""
+    result: list[DataPointCategory] = []
+    if (
+        bool(_CLIMATE_RECEIVER_RE.search(channel_type_name)) is True
+        or channel_type_name in _CLIMATE_SWITCH_RECEIVER_CHANNELS
+    ):
+        result.append(DataPointCategory.CLIMATE)
+    return tuple(result)
+
+
+RECEIVER_PARAMETERS: Final[frozenset[Parameter]] = frozenset(
+    {
+        Parameter.LEVEL,
+        Parameter.STATE,
+    }
+)
+
+
 def check_ignore_parameter_on_initial_load(parameter: str) -> bool:
     """Check if a parameter matches common wildcard patterns."""
     return (